requirements.txt
asgiref==3.3.1
django-environ==0.4.5
django-filter==2.4.0
django==3.1.3
djangorestframework==3.12.2
Pillow==8.0.1
pytz==2020.4
psycopg2==2.8.6